+++ eminus/dft.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: 2021 The eminus developers
+# SPDX-FileCopyrightText: 2022 The eminus developers
 # SPDX-License-Identifier: Apache-2.0
 """Main DFT functions based on the DFT++ formulation."""