+++ examples/07_fod_extra/07_fod_extra.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: 2021 The eminus developers
+# SPDX-FileCopyrightText: 2022 The eminus developers
 # SPDX-License-Identifier: Apache-2.0
 from eminus import Atoms, read, SCF
 from eminus.dft import get_psi