+++ examples/18_k_points/18_k_points.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: 2021 The eminus developers
+# SPDX-FileCopyrightText: 2024 The eminus developers
 # SPDX-License-Identifier: Apache-2.0
 from eminus import Cell, SCF
 from eminus.units import ang2bohr