+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ dev
    - Improve performance of CUBE writing
    - Improve Coulomb potential to use it with different atomic species
    - Fix aspect mode in cell viewer which previously distorted the cell
+   - Fix sampling of non-cubic cells
 
 v3.0.2 - Jan 16, 2025
 ---------------------